Add storybook to input component
